Resolve: Extract `toClientNetwork` (#1719)
* Extract toClientNetwork to shared

* Remove old toClientNetwork

* Move tests
