+++ src/main/api/ledger/thorchain.ts
@@ -1,15 +1,12 @@
 import type Transport from '@ledgerhq/hw-transport'
 import THORChainApp, { LedgerErrorType } from '@thorchain/ledger-thorchain'
-import * as Client from '@xchainjs/xchain-client'
 import { getPrefix } from '@xchainjs/xchain-thorchain'
 import * as E from 'fp-ts/Either'
 
 import { LedgerErrorId, Network } from '../../../shared/api/types'
+import { toClientNetwork } from '../../../shared/utils/client'
 import { getErrorId } from './utils'
 
-// TODO(@Veado) Move `toClientNetwork` from `renderer/services/clients` to `main/util` or so
-const toClientNetwork = (network: Network): Client.Network =>
-  network === 'mainnet' ? Client.Network.Mainnet : Client.Network.Testnet
 // TODO(@veado) Get path by using `xchain-thorchain`
 const PATH = [44, 931, 0, 0, 0]
