+++ src/renderer/services/clients/index.ts
@@ -1,7 +1,6 @@
 export * from './address'
 export * from './common'
 export * from './types'
-export * from './utils'
 export * from './balances'
 export * from './transaction'
 export * from './fees'
