+++ src/shared/utils/client.ts
@@ -1,9 +1,6 @@
 import * as Client from '@xchainjs/xchain-client'
 
-import { Network } from '../../../shared/api/types'
+import { Network } from '../api/types'
 
-/**
- * Helper to type cast `Network` (ASGDX) -> `Client.Network` (xchain-client)
- */
 export const toClientNetwork = (network: Network): Client.Network =>
   network === 'mainnet' ? Client.Network.Mainnet : Client.Network.Testnet
